added automatic publish current song to redis queue in deepcutBot.js & updated discordBot.js
